+++ src/helpers/ContentType.ts
@@ -572,7 +572,7 @@ export class ContentType {
    * @param contentType 
    * @param data 
    */
-  private static async processFields(obj: IContentType | Field, titleValue: string, data: any, filePath: string) {
+  private static async processFields(obj: IContentType | Field, titleValue: string, data: any, filePath: string, isRoot: boolean = true): Promise<any> {
     
     if (obj.fields) {
       const dateFormat = Settings.get(SETTING_DATE_FORMAT) as string;
@@ -581,12 +581,14 @@ export class ContentType {
           if (field.default) {
             data[field.name] = processKnownPlaceholders(field.default, titleValue, dateFormat);
             data[field.name] = await ArticleHelper.processCustomPlaceholders(data[field.name], titleValue, filePath);
-          } else {
+          } else if (isRoot) {
             data[field.name] = titleValue;
+          } else {
+            data[field.name] = ""
           }
         } else {
           if (field.type === "fields") {
-            data[field.name] = await this.processFields(field, titleValue, {}, filePath);
+            data[field.name] = await this.processFields(field, titleValue, {}, filePath, false);
           } else {
             const defaultValue = field.default;